# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: DataFlux API
  description: |
    DataFlux ist eine Universal AI-native Database für Medieninhalte mit einer Plugin-Architektur und Multi-Modal Search Capabilities.
    
    ## Features
    - **Multi-Modal Search**: Suche über Video, Audio, Bilder und Dokumente
    - **AI-Powered Analysis**: Automatische Inhaltsanalyse mit Machine Learning
    - **Real-time Processing**: Streaming und Batch-Verarbeitung
    - **Scalable Architecture**: Microservices-basierte Architektur
    - **Enterprise Security**: JWT Authentication und RBAC
    
    ## Authentication
    Alle API-Endpunkte (außer `/health` und `/docs`) erfordern Authentifizierung über JWT-Token.
    
    ```bash
    # Login
    curl -X POST http://localhost:8006/auth/login \
      -H "Content-Type: application/json" \
      -d '{"username":"admin","password":"admin123"}'
    
    # Use token in requests
    curl -H "Authorization: Bearer <token>" \
      http://localhost:8003/api/v1/search
    ```
  version: 1.0.0
  contact:
    name: DataFlux Support
    email: support@dataflux.com
    url: https://dataflux.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.dataflux.com
    description: Production server
...
paths:
...
  /api/v1/search:
    post:
      tags: [Search]
      summary: Multi-modal search
      description: Search across all media types using text, images, or audio
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Invalid search request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/similar:
    post:
      tags: [Search]
      summary: Find similar content
      description: Find assets similar to a given asset
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                asset_id:
                  type: string
                  format: uuid
                  description: Reference asset ID
                limit:
                  type: integer
                  default: 10
                  minimum: 1
                  maximum: 50
                threshold:
                  type: number
                  default: 0.7
                  minimum: 0.0
                  maximum: 1.0
                  description: Similarity threshold
              required: [asset_id]
      responses:
        '200':
          description: Similar assets
          content:
            application/json:
              schema:
                type: object
                properties:
                  similar_assets:
                    type: array
                    items:
                      $ref: '#/components/schemas/SimilarAsset'

  # Analysis endpoints
...
  /api/v1/stats:
    get:
      tags: [Monitoring]
      summary: System statistics
      description: Get system usage statistics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: System statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SearchRequest:
      type: object
      properties:
        query:
          type: string
          description: Text search query
          example: "cat playing with ball"
        query_type:
          type: string
          enum: [text, image, audio, video]
          default: text
        filters:
          type: object
          properties:
            mime_type:
              type: string
            collection_id:
              type: string
              format: uuid
            date_from:
              type: string
              format: date-time
            date_to:
              type: string
              format: date-time
            min_confidence:
              type: number
              minimum: 0.0
              maximum: 1.0
        limit:
          type: integer
          default: 20
          minimum: 1
          maximum: 100
        offset:
          type: integer
          default: 0
          minimum: 0
      required: [query]
    SearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        total:
          type: integer
        query_time:
          type: number
          description: Query execution time in milliseconds
    SearchResult:
      type: object
      properties:
        asset_id:
          type: string
          format: uuid
        file_name:
          type: string
        mime_type:
          type: string
        similarity_score:
          type: number
          minimum: 0.0
          maximum: 1.0
        thumbnail_path:
          type: string
        metadata:
          type: object
        segments:
          type: array
          items:
            $ref: '#/components/schemas/Segment'
    SimilarAsset:
      type: object
      properties:
        asset_id:
          type: string
          format: uuid
        file_name:
          type: string
        similarity_score:
          type: number
        thumbnail_path:
          type: string
    Segment:
      type: object
      properties:
        segment_id:
          type: string
          format: uuid
        segment_type:
          type: string
        start_time:
          type: number
        end_time:
          type: number
        confidence_score:
          type: number
        segment_data:
          type: object
...
    StatsResponse:
      type: object
      properties:
        total_assets:
          type: integer
        total_collections:
          type: integer
        total_users:
          type: integer
        storage_used:
          type: integer
          format: int64
        processing_queue:
          type: integer
        system_uptime:
          type: integer
          format: int64
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
...